refactor: add signup mutation
